votor: separate out voting_utils
