+++ core/src/replay_stage.rs
@@ -86,10 +86,8 @@ use {
         root_utils,
         vote_history::VoteHistory,
         vote_history_storage::VoteHistoryStorage,
-        voting_loop::{
-            log_leader_change, BLSOp, GenerateVoteTxResult, LeaderWindowNotifier, VotingLoop,
-            VotingLoopConfig,
-        },
+        voting_loop::{log_leader_change, LeaderWindowNotifier, VotingLoop, VotingLoopConfig},
+        voting_utils::{BLSOp, GenerateVoteTxResult},
         CertificateId,
     },
     std::{
